Upgrade studio-player package to 1.3.5
closes ARC-10967
flag=consolidated_media_player

Test Plan:
- Media player works well for direct uploads

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/386326
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Kadir Saglam <kadir.saglam@instructure.com>
Product-Review: Jozsef Gubicza <jozsef.gubicza@instructure.com>

+++ ui/shared/canvas-studio-player/react/CanvasStudioPlayer.tsx
@@ -65,12 +65,16 @@ const convertAndSortMediaSources = (sources: CanvasMediaSource[] | string) => {
   return sources.map(convertMediaSource).sort(byBitrate)
 }
 
+const isCaptionMetaData = (track: MediaTrack | CaptionMetaData): track is CaptionMetaData => {
+  return 'src' in track
+}
+
 const convertMediaTracksIfNeeded = (
   tracks: MediaTrack[] | CaptionMetaData[],
 ): CaptionMetaData[] => {
   // @ts-expect-error
   return tracks.map(track => {
-    if ('src' in track) return track
+    if (isCaptionMetaData(track)) return track
     return {
       locale: track.locale,
       language: captionLanguageForLocale(track.locale),
@@ -240,12 +244,14 @@ export default function CanvasStudioPlayer({
       return
     }
 
-    await fetch(caption.src, {
-      method: 'DELETE',
-      headers: {
-        'X-CSRF-Token': getCookie('_csrf_token'),
-      },
-    })
+    if (caption.src) {
+      await fetch(caption.src, {
+        method: 'DELETE',
+        headers: {
+          'X-CSRF-Token': getCookie('_csrf_token'),
+        },
+      })
+    }
     setMediaCaptions(prev => prev?.filter(c => c.src !== caption.src))
   }, [])
